fdj philo pas mal

+++ Philo_src/_Include/philo.h
@@ -6,7 +6,7 @@
 /*   By: ngriveau <ngriveau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/23 16:05:10 by pbizien           #+#    #+#             */
-/*   Updated: 2023/03/29 18:13:34 by ngriveau         ###   ########.fr       */
+/*   Updated: 2023/03/29 19:04:16 by ngriveau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ typedef struct s_human
 {
 	int				nb;
 	int				status;
-	int				timing;
+	long			timing;
 	int				*leftfork;
 	pthread_mutex_t	*leftmutex;
 	int				*rightfork;
